place api in respective controller, abstract from routes and server.js
